+++ plugins/newspack-plugin/CHANGELOG.md
@@ -1,3 +1,10 @@
+## newspack [6.44.4](https://github.com/Automattic/newspack-workspace/compare/newspack@6.44.3...newspack@6.44.4) (2026-07-02)
+
+
+### Bug Fixes
+
+* **content-gate:** keep commenting per Discussion Settings on metered posts (NPPD-1829, [#459](https://github.com/Automattic/newspack-workspace/issues/459)) ([43d4fc6](https://github.com/Automattic/newspack-workspace/commit/43d4fc64543b8e8be5ba21529ac6eb1194fbe0af))
+
 ## newspack [6.44.3](https://github.com/Automattic/newspack-workspace/compare/newspack@6.44.2...newspack@6.44.3) (2026-07-01)